Change the layout, add more filler text
Changed the layout to make the website look cleaner. Decreased the size
of my name, so it looks less vainful. Added all the navigation buttons
to the right of my name. Put the date on top of the title. Changed the
color of the tags. Change the color of text everywhere, changed the font
size, font-family etc.

Added more lorem-ipsum filler text to sample blog posts.