Add an aboutMe page
When the name Aditya Ambadipudi is clicked or when the top level route
is open, open the aboutMe page. For this, add an MDX file called
aboutMe.mdx. Add a function in the mdxUtil file for reading and parsing
this MDX file. In the top level route use this utility function to parse
and display the aboutme mdx file.