Add the chess puzzles page
Add a method to the mdxUtil file to read and parse the chess puzzles mdx
files. Add a class called ChessPuzzle to the models directory that will
be used by the aforementioned mdxUtil function.

Add a component for displaying individual puzzles. Add a component for
displaying a collection of puzzles. Add another component that will serve
as a button for displaying hints and solutions. Only when the user clicks
on this button will the hidden text be displayed.

Finally add a route called chess puzzles that will use these components
and display the stored chess puzzles.